### 2016-06-24 #32 @whedon
Submission: Advanced viewshed analysis: a Quantum GIS plug-in for the analysis of visual landscapes
**Submitting author:** @zoran-cuckovic (<a href="http://orcid.org/0000-0001-7626-4086">Zoran Čučković</a>)
**Repository:** <a href="https://github.com/zoran-cuckovic/QGIS-visibility-analysis" target ="_blank">https://github.com/zoran-cuckovic/QGIS-visibility-analysis</a>
**Version:** 0.5.1
