### 2016-06-24 75ad923
Update paper.md
