### 2016-06-30 #32 @zoran-cuckovic
Hello!

To test my software all you need is to install the plugin into QGIS from the official repository (as any other standard plugin). The current version (submitted here ) has some 1000 dowlands - without particular errors reported.
