### 2016-07-06 #32 @labarba
@zoran-cuckovic — Can you recommend any distribution list where we may ask for help reviewing your submission? I made a call via Twitter:
https://twitter.com/lorenaabarba/status/748864879902195712
... which got retweeted by The Open Source Geospatial Foundation to their 13.9K followers. Yet no one has come forward.
I also targeted someone, but he said it is not his expertise:
https://twitter.com/lorenaabarba/status/748895404578701312

And today, I asked for help with
https://twitter.com/lorenaabarba/status/750817953440997376

Not sure how  else to reach out!
