### 2016-07-07 #32 @zoran-cuckovic
@labarba, @arfon  I have made a request over QGIS developer mailing list.
Otherwise, my colleagues are mostly human sciences and not software developers - what would be required for a reviewer? Developing software, publishing on software use, actively collaborating over git-hub or similar platform?
