### 2016-07-07 #32 @labarba
I also got some response via Twitter:
https://twitter.com/geomenke/status/750878480707923968
