### 2016-07-07 #32 @volaya
I don't have enough time for a full review, but just a quick comment:

The algorithm looks useful and interesting, but, being part of QGIS, it makes little sense to have it outside of the Processing framework. It would be much more useful and powerful if integrated into Processing.

Hope this helps
