### 2016-07-07 #32 @giohappy
I must agree with Victor. Looking at the algorithm configuration, it would
fit very well within the Processing toolbox, but I imagine this could be a
future refactoring/enhancement...

2016-07-07 12:44 GMT+02:00 volaya <notifications@github.com>:

> I don't have enough time for a full review, but just a quick comment:
>
> The algorithm looks useful and interesting, but, being part of QGIS, it
> makes little sense to have it outside of the Processing framework. It would
> be much more useful and powerful if integrated into Processing.
>
> Hope this helps
>
