### 2016-07-07 #32 @kyleniemeyer
Robin Wilson (@robintw) volunteered to review via Twitter!

@arfon: can you add him as a JOSS reviewer?
