### 2016-07-07 #32 @zoran-cuckovic
@volaya , @giohappy
I'm not sure the actual relationship of my software with the QGIS core is relevant here (?). I prefer having the plugin solution because inside there are several algorithms combined together and there are many options - and because I plan further enhancements. There are many visibility indices that can be added in the future, so the plan (optimistic) is to create some kind of a toolbox... But, yes, I understand that using processing toolbox allows to combine my algorithm with others: I'm planning to make simplified version(s) for that purpose.
