### 2016-07-12 #32 @arfon
> Robin Wilson (@robintw) volunteered to review via Twitter!

Fantastic! @robintw - I've invited you to `@openjournals/joss-reviewers` team which should allow you to update this issue with your review. Also, please take a look at the reviewer guidelines to familiarized yourself with the process http://joss.theoj.org/about#reviewer_guidelines
