### 2016-07-23 #32 @robintw
Sorry for the delay in dealing with this - I'm getting on with it now. Am I meant to actually tick the boxes in the checklist in the first post - and if so, how? I don't seem to be able to tick them - should I copy the checklist to reply and tick them there?
