### 2016-07-23 #32 @arfon
> Am I meant to actually tick the boxes in the checklist in the first post - and if so, how? I don't seem to be able to tick them - should I copy the checklist to reply and tick them there?

Yes please @robintw. You should have received an invite to the `@openjournals/joss-reviewers` team which will give you appropriate permissions. You can accept the invitation by clicking the link at the top of the page here: https://github.com/openjournals
