### 2016-07-23 #32 @robintw
Thanks @arfon - I somehow managed to miss that email. I've now got permissions to edit the checkboxes. I'll post the specific review comments shortly.
