### 2016-07-23 #32 @robintw
In general, the software seems to work ok, but there are various issues that mean that I can't tick all of the boxes in the review checklist. I'll go through these points individually below in the checklist categories:

**General:**
- **Versions**
  - The latest version is 0.5.1, and this is what is installed from the QGIS Plugin Manager, but https://hub.qgis.org/projects/viewshed lists latest version as 0.4.1. Please clarify/update as necessary.

**Functionality:**
- **Installation**
   - There are no installation instructions. I knew to use the QGIS Plugin Repository, but not all potential users will know this. Please add basic instructions.
  -  Can this plugin be installed manually (ie. from the cloned git repository rather than through the QGIS Plugin Manager) - if so, please provide instructions. This can be particularly useful when installing on computers without an internet connection.
  - There is no information on what versions of QGIS this plugin works with - please add.
- **Functionality**
  - The plugin seems to work ok with the tests that I've done. However, I have no idea whether it is producing the correct results because of the lack of specific test cases (see below)
- **Performance**
  - The plugin seems to run fast in my limited testing of it. However, there are no large test inputs provided and no speed comparisons with other tools in the documentation, so I am unable to properly comment on this.
   - The author claims that the latest version of this tool (v0.5.1) is significantly faster than the previous versions. However, I could not confirm this as I couldn't find any previous versions (there are no previous releases listed on Github, no tags for previous versions, etc).

**Documentation:**
- **General**
  - Generally the documentation has quite a lot of information on the underlying algorithms etc, but is fairly brief on the actual use of the software, and doesn't contain anything like useful screenshots.
  - The README is very short and has no information on dependencies or installation - both of which are key for a README. The README and documentation should also mention the license, but they don't.
- **Installation instructions:**
  - None provided (see above). Please add.
- **Example usage**
  - There is a folder called `TEST_DATA` with some example inputs, but no description of what any of the inputs are, or what outputs should be produced using these inputs and certain parameter configurations. A 'tutorial' section in the documentation going through an example analysis with data from TEST_DATA, and showing the 'correct' output would be very useful. Output files for the correct output with a range of different inputs and parameter settings should also be provided to allow users to test that the software is functioning as intended.
  - There are no screenshots in any of the documentation, and in fact, no information on where to find the menu items within the QGIS interface (not all Plugins create menu entries under the Plugins menu!)
- **Functionality documentation**:
  - Most parameters are documented in the 'reference' tab of the user-interface - but some are missing. These should be added.
  - The code itself is not well documented: there are no docstrings, and there are large chunks of code commented out (eg. https://github.com/zoran-cuckovic/QGIS-visibility-analysis/blob/master/viewshedanalysis.py#L205) and comments like `#open console !!!! to monitor !!! #and test time !!!!` or `NOT WORKING`, which do not give confidence!
- **Automated tests**
  - None are present. Ideally these should be added, but at the very least examples of correct outputs for multiple sets of input parameters should be provided.
- **Community Guidelines**:
  - None are present. Please add.

**Paper:**
The paper is generally ok, except there are no DOIs in the references, even for the journal papers that should have a DOI. I would also suggest referencing the peer-reviewed paper on GRASS (which has a DOI), rather than just its manual - see https://grasswiki.osgeo.org/wiki/GRASS_Citation_Repository

**General:**
I agree with the other comments, that this plugin would be far more useful if it were integrated into the processing toolbox - almost all actual geoprocessing tools in QGIS are being integrated into the processing toolbox. I'm not sure why you suggested making a simpler plugin for use within the toolbox: processing tools can have just as many options, parameters and so on as you can in your own plugin! In fact, the latest version of QGIS includes far better support for releasing processing scripts as parts of a Python plugin. Although integration with the processing toolbox would not be a requirement for this publication, I would strongly suggest it as a sensible way forward.
