### 2016-07-24 #32 @arfon
👍 thanks for this thoughtful and thorough review @robintw.

@zoran-cuckovic - please update this thread when you've had a chance to review @robintw's feedback and make the necessary changes.
