### 2016-07-27 6fbfe3d
Documentation fix and test data update
