### 2016-07-29 dce44df
Add DEM raster etc
