+++ test_data/Reference output/Parameters.md
@@ -1,12 +1,0 @@
-Parameters
-==========
-
-Observer point: single point.shp
-Elevation data: DEM.tiff
-
-Settings
---------
-Radius of analysis: 15 000
-Observer height: 1.6 (default)
-Earth curvature: Yes
-refraction: default
