### 2016-07-29 26f975a
Add DEM, observer points etc
