+++ README.md
@@ -1,0 +1,8 @@
+Test data
+=========
+
+This folder contains shapefiles with different observation points and an extract of SRTM DEM for tutorials and for testing purposes. The "reference output" folder contains some analysis outputs for comparison.
+
+Projection system: EPSG:31275 (MGI / Balkans zone 5).
+
+For tutorials check [zoran-cuckovic.from.hr](zoran-cuckovic.from.hr/landscape_analysis/visibility) 
