### 2016-07-29 1bf7457
Add cumulative output
