+++ ui_viewshedanalysis.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'ui_viewshedanalysis.ui'
 #
-# Created: Sat Apr 23 16:44:57 2016
+# Created: Fri Jul 29 17:27:20 2016
 #      by: PyQt4 UI code generator 4.10.2
 #
 # WARNING! All changes made in this file will be lost!
@@ -28,7 +28,8 @@ class Ui_ViewshedAnalysis(object):
         ViewshedAnalysis.setObjectName(_fromUtf8("ViewshedAnalysis"))
         ViewshedAnalysis.resize(448, 622)
         self.tabWidget = QtGui.QTabWidget(ViewshedAnalysis)
-        self.tabWidget.setGeometry(QtCore.QRect(10, 10, 427, 571))
+        self.tabWidget.setGeometry(QtCore.QRect(10, 10, 427, 581))
+        self.tabWidget.setAutoFillBackground(False)
         self.tabWidget.setObjectName(_fromUtf8("tabWidget"))
         self.General_tab = QtGui.QWidget()
         self.General_tab.setObjectName(_fromUtf8("General_tab"))
@@ -55,7 +56,7 @@ class Ui_ViewshedAnalysis(object):
         self.cmdBrowse.setGeometry(QtCore.QRect(320, 90, 75, 31))
         self.cmdBrowse.setObjectName(_fromUtf8("cmdBrowse"))
         self.Output_groupBox = QtGui.QGroupBox(self.General_tab)
-        self.Output_groupBox.setGeometry(QtCore.QRect(10, 350, 391, 91))
+        self.Output_groupBox.setGeometry(QtCore.QRect(10, 340, 391, 91))
         self.Output_groupBox.setObjectName(_fromUtf8("Output_groupBox"))
         self.chkBinary = QtGui.QRadioButton(self.Output_groupBox)
         self.chkBinary.setGeometry(QtCore.QRect(10, 30, 141, 22))
@@ -120,7 +121,7 @@ class Ui_ViewshedAnalysis(object):
         self.label_8.setGeometry(QtCore.QRect(10, 30, 91, 16))
         self.label_8.setObjectName(_fromUtf8("label_8"))
         self.groupBox = QtGui.QGroupBox(self.General_tab)
-        self.groupBox.setGeometry(QtCore.QRect(10, 450, 391, 81))
+        self.groupBox.setGeometry(QtCore.QRect(10, 440, 391, 111))
         self.groupBox.setObjectName(_fromUtf8("groupBox"))
         self.chkCurvature = QtGui.QCheckBox(self.groupBox)
         self.chkCurvature.setGeometry(QtCore.QRect(20, 50, 161, 22))
@@ -136,6 +137,16 @@ class Ui_ViewshedAnalysis(object):
         self.chkCumulative.setEnabled(True)
         self.chkCumulative.setGeometry(QtCore.QRect(20, 20, 261, 21))
         self.chkCumulative.setObjectName(_fromUtf8("chkCumulative"))
+        self.label_12 = QtGui.QLabel(self.groupBox)
+        self.label_12.setGeometry(QtCore.QRect(20, 80, 46, 13))
+        self.label_12.setObjectName(_fromUtf8("label_12"))
+        self.cmbAlgorithm = QtGui.QComboBox(self.groupBox)
+        self.cmbAlgorithm.setGeometry(QtCore.QRect(190, 80, 121, 22))
+        self.cmbAlgorithm.setEditable(False)
+        self.cmbAlgorithm.setObjectName(_fromUtf8("cmbAlgorithm"))
+        self.cmbAlgorithm.addItem(_fromUtf8(""))
[…]
