+++ doViewshed.py
@@ -13,18 +13,11 @@ email : /
 * *
 * This program is free software; you can redistribute it and/or modify *
 * it under the terms of the GNU General Public License as published by *
-* the Free Software Foundation; either version 2 of the License, or *
-* (at your option) any later version. *
+* the Free Software Foundation version 2 of the License, or *
+* any later version. *
 * *
 ***************************************************************************/
 """
-
-#open console !!!! to monitor !!!
-#and test time !!!!
-
-
-
-
 
 
 from __future__ import division 
@@ -51,26 +44,19 @@ def dist(x1,y1,x2,y2, estimation=False):
 
     return r
 
-def sum_by_group(values, groups):
-    #http://stackoverflow.com/questions/4373631/sum-array-by-number-in-numpy
-    order = numpy.argsort(groups)
-    groups = groups[order]
-    values = values[order]
-    values.cumsum(out=values)
-    index = numpy.ones(len(groups), 'bool')
-    index[:-1] = groups[1:] != groups[:-1]
-    values = values[index]
-    groups = groups[index]
-    values[1:] = values[1:] - values[:-1]
-    return values, groups
 
 def error_matrix(radius, size_factor=1):
 
+    """
+    Create a set of lines of sight which can be reused for all calculations. 
+    Each line (departing from the observer point) has its target and a set of pixels it passes through.
+    Only 1/8th of full radius is enough : the rest can be copied/mirrored. 
+    """
+
     
-    
-    radius_large =  radius + radius * (size_factor-1)  #if double else 0
+    if size_factor == 0: size_factor = 1 #0 is for coarse algo...
+    radius_large = radius  * size_factor  #if double else 0  
                                                 
-    # SMANJI JEDNU DIMENZIJU
     mx_index= numpy.zeros((radius_large +1 , radius, 4))
 
     min_err = {}
@@ -96,9 +82,7 @@ def error_matrix(radius, size_factor=1):
             else:
                 y_f += 1
                 D += dy - dx
-##          # it is not necessary to make a circle, all calculations are in a square matrix (when x / y are not specified, they are 0,0)
-##            if x_f + y_f > 2*radius : #make a circle
-##                if dist (radius, radius, x_f, y_f) > radius +0.4999:break # +0.5 = centre of pix
+#        # it is not necessary to make a circle, all calculations are in a square matrix (when x / y are not specified, they are 0,0)
                            
             #reverse x,y for data array!
             yx= (y_f,x_f)
@@ -126,24 +110,29 @@ def error_matrix(radius, size_factor=1):
 
     return mx_index
     
-   
-
-
-
 
 def Viewshed (Obs_points_layer, Raster_layer, z_obs, z_target, radius, output,
               output_options,
               Target_layer=0, search_top_obs=0, search_top_target=0,
               z_obs_field=0, z_target_field=0,
-              curvature=0, refraction=0): #for visib index !!
-# ########################################################
-#    output_options[0]= "Fast"
-    fast = False #dodatak !
-##########################################################
-        
-    def search_top_z (pt_x, pt_y, search_top): #shoud be a separate loop for all points ?
+              curvature=0, refraction=0, algorithm = 1): 
 
-        # global variable      
+
+    """
+    Opens a DEM and produces viewsheds from a set of points. Algorithms for raster ouput are all based on a same approach,
+    explained at zoran-cuckovic.from.hr (perhaps published one day...). Intervisibility calculation, however,
+    is on point to point basis and has its own algorithm. 
+    """
+
+    fast = False #non-interpolated algorithm for next version
+
+    def search_top_z (pt_x, pt_y, search_top):
+        """
+        Find the highest point in a perimeter around each observer point.
[…]
