+++ metadata.txt
@@ -14,7 +14,7 @@ category=Raster
 name=Viewshed Analysis
 qgisMinimumVersion=2.0
 description=Advanced visibility analysis tool.
-version=0.5.1
+version=0.5.2
 author=Zoran Čučković
 email=cuckovic.zoran@google.com
 about=Viewshed analysis calculates visible surface from a given observer point over a digital elevation model. This plugin is native to QGIS and intended for more complex modelling, such as the depth below the visible horizon or generation of intervisibilty networks between groups of points. It is particularly performant for multiple viewshed calculations form a set of fixed points. For single point, large area viewsheds the algorithms already available in QGIS (GRASS r.los or SAGA viewshed module) would do the job. 
@@ -36,7 +36,7 @@ tracker=https://github.com/zoran-cuckovic/QGIS-visibility-analysis/issues
 repository=https://github.com/zoran-cuckovic/QGIS-visibility-analysis
 icon=./icon.png
 # experimental flag
-experimental=True
+experimental=False
 
 # deprecated flag (applies to the whole plugin, not just a single version
 deprecated=False

[…]
