### 2016-07-29 702d445
Update ui_viewshedanalysis.py
