### 2016-07-29 0b6b210
Update paper.md
