+++ Reference output/Parameters.md
@@ -23,7 +23,8 @@ Elevation data: DEM.tiff
 Radius of analysis: 5 000  
 Observer height: 0  
 Target height: 1.6  
-Find higher point: 1 pixel   
+Find higher point: 1 pixel  
+Output type: binary viewshed
 Earth curvature: Yes  
 refraction: default  
 
