### 2016-07-29 #32 @zoran-cuckovic
Hello!
As there were quite a number of things to improve, including the documentation in GUI, I have made a new release – it has to be installed manually, as described in [the manual](http://zoran-cuckovic.github.io/QGIS-visibility-analysis/) (with a download link).

*Functionality:*
1) Installation
-  There are no installation instructions. **Fixed** in the manual (link above)
-  Information on what versions of QGIS this plugin works with. **Fixed**.

2) Functionality
- Lack of specific test cases. **Fixed**: see the [branch with test data]( https://github.com/zoran-cuckovic/QGIS-visibility-analysis/tree/test-data) .

3) Performance
- There are no large test inputs provided and no speed comparisons with other tools in the documentation. **Fixed** : data and parameters used in the above link, plus a [speed comparison here]( http://zoran-cuckovic.from.hr/qgis-visibility-analysis-algorithm/) (hope the test would suffice…)
- I couldn't find any previous versions (no previous releases listed on Github, etc). >> I'm not a proficient GitHub user. All previous versions can be found in [QGIS plugins repo](https://plugins.qgis.org/plugins/ViewshedAnalysis/)

*Documentation:*
1) General
- The documentation (…) is fairly brief on the actual use of the software, and doesn't contain anything like useful screenshots. **Fixed**: [a tutorial made]( http://zoran-cuckovic.from.hr/qgis-viewshed-plugin-a-tutorial/)
- The README is very short and has no information on dependencies or installation. **Fixed**

2) Installation instructions:
- None provided **Fixed** (see above).

3) Example usage
- A 'tutorial' section in the documentation going through an example analysis (…) **Fixed**. Output files for the correct output with a range of different inputs and parameter settings should also be provided (…)**Fixed** (see above "functionality").
- There are no screenshots in any of the documentation, and in fact, no information on where to find the menu items within the QGIS interface. **Fixed** in the manual.

4) Functionality documentation:
- Most parameters are documented in the 'reference' tab of the user-interface - but some are missing. These should be added. **Fixed**
- The code itself is not well documented: there are no docstrings, and there are large chunks of code commented out. **Fixed** (?): I've cleaned the code and added docstrings at beginnings of functions.

5) Automated tests
- None are present. Ideally these should be added, but at the very least examples of correct outputs for multiple sets of input parameters should be provided. **Fixed** (sort of ?) I have added a git branch containing test data, desired output and information on parameters used (5 outputs in all). Same data is used in the tutorial I made recently (link above).

6) Community Guidelines:
- None are present. Please add. **Fixed** in Readme + statement added.

*Paper:*
The paper is generally ok, except there are no DOIs in the references **Fixed**. I would also suggest referencing the peer-reviewed paper on GRASS **Fixed**.

**Thanks a lot to the reviewer !**
