### 2016-07-30 #32 @arfon
@zoran-cuckovic - thanks for these updates.

@robintw - can when you get a chance could you review these updates and let me know if we're good to accept this submission now?
