### 2016-08-05 #32 @arfon
Thanks @robintw!

@zoran-cuckovic - could you make an archive of the current release in Zenodo/figshare and copy the DOI into this thread?
