### 2016-08-07 #32 @lidapeng
Hello Arfon @arfon . Is the review process for this paper done? If not I want to be a reviewer for this paper. Do I need to register a reviewer account?
I am running the plugin using the test data given in the data branch. The cumulative viewshed analysis takes a very long time. It also generates many files in the directory. Should I manually install the latest version? @zoran-cuckovic   Thanks.
![viewshed](https://cloud.githubusercontent.com/assets/3634010/17465087/80d04c3a-5cbb-11e6-9745-41350a5006a5.JPG)
