### 2016-08-08 #32 @arfon
> Hello Arfon @arfon . Is the review process for this paper done? If not I want to be a reviewer for this paper. Do I need to register a reviewer account?

@lidapeng - thanks for your interest. This review is now complete. If you're looking for support using the package I'd recommend you take a look at the project repository here: https://github.com/zoran-cuckovic/QGIS-visibility-analysis
