### 2016-08-09 #32 @zoran-cuckovic
@lidapeng : you shoud select "cumulative" option. If it still doesn't work for you, please raise an issue in the repository.
