### 2016-08-09 #32 @zoran-cuckovic
@arfon, @robintw
I have deposited the submitted version in Zenodo: [http://dx.doi.org/10.5281/zenodo.59896](http://dx.doi.org/10.5281/zenodo.59896)
The badge : [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.59896.svg)](http://dx.doi.org/10.5281/zenodo.59896)

Thanks again JOSS and reviewer!
