### 2016-08-09 #32 @arfon
@zoran-cuckovic - apologies for this request coming so late in the day but would you mind moving the references into the `paper.bib` file and into a bibtex format? We can't do the final paper processing/reference management with Crossref without this sorry.
