### 2016-08-09 d6b139d
Create paper.bib
