+++ paper/paper.bib
@@ -1,0 +1,59 @@
+@INPROCEEDINGS {Cuckovic2015,
+    author    = "Čučković, Zoran",
+    title     = "Exploring intervisibility networks: a case study from Bronze and Iron Age Istria (Croatia and Slovenia).",
+    booktitle = "CAA 2014 - 21st Century Archaeology : Proceedings of the 42nd Annual Conference on Computer Applications and Quantitative Methods in Archaeology",
+    year      = "2015",
+    editor    = "Giligny, François and Djindjian, François and Costa, Laurent and Moscati, Paola and Robert, Sandrine",
+    pages     = "469 - 478.",
+    address   = "Paris, France"
+}
+
+@MISC {ERDAS2015,
+    author       = "ERDAS",
+    title        = "ERDAS Imagine Suite 2015",
+    publisher = "Norcross: Hexagon Geospatial",
+    year         = "2015"
+}
+
+@MISC {ESRI2016,
+    author       = "ESRI",
+    title        = "ArcGis Desktop: Release 10.4",
+    publisher = "Redlands: Environmental Systems Research Institute.",
+    year         = "2016"
+}
+
+@BOOK {HIGUCHI1983,
+    author    = "Higuchi, Tadahiko",
+    title     = "Visual and Spatial Structure of Landscapes.",
+    publisher = "MIT Press",
+    year      = "1983",
+    address   = "Cambridge, Mass."
+}
+
+@ARTICLE{Llobera2003,
+author = {  Llobera,  Marcos  },
+title = {Extending GIS-based visual analysis: the concept of visualscapes},
+journal = {International Journal of Geographical Information Science},
+volume = {17},
+number = {1},
+pages = {25-48},
+year = {2003},
+doi = {10.1080/713811741},
+URL = {http://dx.doi.org/10.1080/713811741},
+eprint = {http://dx.doi.org/10.1080/713811741}
+}
+
+@ARTICLE{Neteler2012,
+author = "Neteler, Markus and Hamish Bowman, M. and Landa, Martin  and Metz, Markus"
+title = "GRASS GIS: A multi-purpose open source GIS ",
+journal = "Environmental Modelling & Software ",
+volume = "31",
+number = "",
+pages = "124 - 130",
+year = "2012",
+note = "",
+issn = "1364-8152",
+doi = "http://dx.doi.org/10.1016/j.envsoft.2011.11.014",
+url = "http://www.sciencedirect.com/science/article/pii/S1364815211002775",
+
+}
