+++ paper/paper.md
@@ -15,7 +15,7 @@ bibliography: paper.bib
 
 # Summary
 
-Viewshed analysis is a standard feature of GIS software packages, such as ArcGIS (ESRI 2016), GRASS (Neteler *et al.* 2015) or ERDAS (2015). However, these implementations vary considerably in terms of their versatility and robustness. Software in the free domain is particularly poor in this respect: visibility analysis is generally implemented as a simple binary query (true/false) for elevation datasets (eg. GRASS or SAGA GIS). In order to meet the demands of a typical analysis concerning visual landscapes we would be interested to find out how deep are certain locations below the visible horizon, what is the overall visual potential of a landscape or which sites are connected in visual networks (cf. Higouchi 1983; Llobera 2003; Čučković 2015).
+Viewshed analysis is a standard feature of GIS software packages, such as ArcGIS (ESRI 2016), GRASS (Neteler *et al.* 2015) or ERDAS (2015). However, these implementations vary considerably in terms of their versatility and robustness. Software in the free domain is particularly poor in this respect: visibility analysis is generally implemented as a simple binary query (true/false) for elevation datasets (eg. GRASS or SAGA GIS). In order to meet the demands of a typical analysis concerning visual landscapes we would be interested to find out how deep are certain locations below the visible horizon, what is the overall visual potential of a landscape or which sites are connected in visual networks (cf. Higuchi 1983; Llobera 2003; Čučković 2015).
 
 Advanced viewshed analysis plug-in for open source Quantum GIS software has been made in order to meet some of these demands: besides standard binary viewshed it provides information on the depth at which objects may be hidden from view, mapping of visual horizon and analysis of intervisibility networks. As of version 0.5.1, the implemented algorithm has been adapted for intensive, repetitive viewshed calculation from multiple observation points. 
 
@@ -31,7 +31,7 @@ ERDAS 2015. ERDAS Imagine Suite 2015. Norcross: Hexagon Geospatial.
 
 ESRI 2016. ArcGIS Desktop: Release 10.4. Redlands: Environmental Systems Research Institute.
 
-Higouchi, T. 1983. *Visual and Spatial Structure of Landscapes*. Cambridge, Mass: MIT Press. 
+Higuchi, T. 1983. *Visual and Spatial Structure of Landscapes*. Cambridge, Mass: MIT Press. 
 
 Llobera, M. 2003. Extending GIS-based visual analysis: the concept of visualscapes. International journal of geographical information science 17(1), 25-48. doi: [10.1080/713811741](https://doi.org/10.1080/713811741)
 
