+++ paper/paper.bib
@@ -31,16 +31,16 @@
 }
 
 @ARTICLE{Llobera2003,
-author = {  Llobera,  Marcos  },
-title = {Extending GIS-based visual analysis: the concept of visualscapes},
-journal = {International Journal of Geographical Information Science},
-volume = {17},
-number = {1},
-pages = {25-48},
-year = {2003},
-doi = {10.1080/713811741},
-URL = {http://dx.doi.org/10.1080/713811741},
-eprint = {http://dx.doi.org/10.1080/713811741}
+author = " Llobera,  Marcos  ",
+title = "Extending GIS-based visual analysis: the concept of visualscapes",
+journal = "International Journal of Geographical Information Science",
+volume = "17",
+number = "1",
+pages = "25-48",
+year = "2003",
+doi = "10.1080/713811741",
+URL = "http://dx.doi.org/10.1080/713811741",
+eprint = "http://dx.doi.org/10.1080/713811741"
 }
 
 @ARTICLE{Neteler2012,
