### 2016-08-09 80e52f3
Update paper.bib
