+++ paper/paper.bib
@@ -5,21 +5,24 @@
     year      = "2015",
     editor    = "Giligny, François and Djindjian, François and Costa, Laurent and Moscati, Paola and Robert, Sandrine",
     pages     = "469 - 478.",
+    publisher = "Archaeopress"
     address   = "Paris, France"
 }
 
 @MISC {ERDAS2015,
     author       = "ERDAS",
     title        = "ERDAS Imagine Suite 2015",
-    publisher = "Norcross: Hexagon Geospatial",
+    publisher = "Hexagon Geospatial",
+    address    = "Norcross",
     year         = "2015"
 }
 
 @MISC {ESRI2016,
     author       = "ESRI",
     title        = "ArcGis Desktop: Release 10.4",
-    publisher = "Redlands: Environmental Systems Research Institute.",
-    year         = "2016"
+    publisher = "Environmental Systems Research Institute.",
+    year         = "2016",
+    address      = "Redlands"
 }
 
 @BOOK {HIGUCHI1983,
