### 2016-08-09 5cba03f
Removed refs
