### 2016-08-09 #32 @zoran-cuckovic
@arfon - I've created the .bib file. Hope it works!
