### 2016-08-09 #32 @arfon
@zoran-cuckovic - your paper is now accepted into JOSS. The DOI is http://dx.doi.org/10.21105/joss.00032 🎉 🚀 💥

Thanks for the review @robintw!
